+++ lte/components/model-table/footer.blade.php
@@ -3,14 +3,14 @@
         <div class="col-sm d-none d-lg-block d-xl-block">
             <div class="btn-group justify-content-start" role="group">
                 @foreach($per_pages as $per)
-                    <a href="{{urlWithGet([$per_name => $per], [$page_name])}}"
+                    <a href="{{admin_url_with_get([$per_name => $per], [$page_name])}}"
                        class="btn btn-sm btn-{{$per == $per_page ? 'dark' : 'secondary'}}">{{$per}}</a>
                 @endforeach
             </div>
         </div>
         <div class="col-sm">
             <div
-                style="text-align: center">{{__('admin.showing_to_of_entries', ['show' => $from, 'to' => $to, 'of' => $paginator->total()])}}</div>
+                    style="text-align: center">{{__('admin.showing_to_of_entries', ['show' => $from, 'to' => $to, 'of' => $paginator->total()])}}</div>
         </div>
         <div class="col-sm d-none d-lg-block d-xl-block">
             @if ($paginator->hasPages())
@@ -18,12 +18,14 @@
                     <ul class="pagination justify-content-end pagination-sm mb-0">
                         {{-- Previous Page Link --}}
                         @if ($paginator->onFirstPage())
-                            <li class="page-item disabled" aria-disabled="true" aria-label="@lang('pagination.previous')">
+                            <li class="page-item disabled" aria-disabled="true"
+                                aria-label="@lang('pagination.previous')">
                                 <span class="page-link" aria-hidden="true">&lsaquo;</span>
                             </li>
                         @else
                             <li class="page-item">
-                                <a class="page-link" href="{{ urlWithGet([$page_name => $paginator->currentPage() - 1]) }}"
+                                <a class="page-link"
+                                   href="{{ admin_url_with_get([$page_name => $paginator->currentPage() - 1]) }}"
                                    rel="prev" aria-label="@lang('pagination.previous')">&lsaquo;</a>
                             </li>
                         @endif
@@ -33,7 +35,7 @@
                             {{-- "Three Dots" Separator --}}
                             @if (is_string($element))
                                 <li class="page-item disabled" aria-disabled="true"><span
-                                        class="page-link">{{ $element }}</span></li>
+                                            class="page-link">{{ $element }}</span></li>
                             @endif
 
                             {{-- Array Of Links --}}
@@ -41,10 +43,10 @@
                                 @foreach ($element as $page => $url)
                                     @if ($page == $paginator->currentPage())
                                         <li class="page-item active" aria-current="page"><span
-                                                class="page-link">{{ $page }}</span></li>
+                                                    class="page-link">{{ $page }}</span></li>
                                     @else
                                         <li class="page-item"><a class="page-link"
-                                                                 href="{{ urlWithGet([$page_name => $page]) }}">{{ $page }}</a>
+                                                                 href="{{ admin_url_with_get([$page_name => $page]) }}">{{ $page }}</a>
                                         </li>
                                     @endif
                                 @endforeach
@@ -54,7 +56,8 @@
                         {{-- Next Page Link --}}
                         @if ($paginator->hasMorePages())
                             <li class="page-item">
-                                <a class="page-link" href="{{ urlWithGet([$page_name => $paginator->currentPage() + 1]) }}"
+                                <a class="page-link"
+                                   href="{{ admin_url_with_get([$page_name => $paginator->currentPage() + 1]) }}"
                                    rel="next" aria-label="@lang('pagination.next')">&rsaquo;</a>
                             </li>
                         @else
@@ -80,7 +83,8 @@
                         </li>
                     @else
                         <li class="page-item">
-                            <a class="page-link" href="{{ urlWithGet([$page_name => $paginator->currentPage() - 1]) }}"
+                            <a class="page-link"
+                               href="{{ admin_url_with_get([$page_name => $paginator->currentPage() - 1]) }}"
                                rel="prev">@lang('pagination.previous')</a>
                         </li>
                     @endif
@@ -88,7 +92,8 @@
                     {{-- Next Page Link --}}
                     @if ($paginator->hasMorePages())
                         <li class="page-item">
-                            <a class="page-link" href="{{ urlWithGet([$page_name => $paginator->currentPage() + 1]) }}"
+                            <a class="page-link"
+                               href="{{ admin_url_with_get([$page_name => $paginator->currentPage() + 1]) }}"
                                rel="next">@lang('pagination.next')</a>
                         </li>
                     @else