+++ lte/components/time-line/paginate-footer.blade.php
@@ -4,14 +4,14 @@
             <div class="col-sm d-none d-lg-block d-xl-block">
                 <div class="btn-group justify-content-start" role="group">
                     @foreach($per_pages as $per)
-                        <a href="{{urlWithGet([$per_name => $per], [$page_name])}}"
+                        <a href="{{admin_url_with_get([$per_name => $per], [$page_name])}}"
                            class="btn btn-sm btn-{{$per == $per_page ? 'dark' : 'secondary'}}">{{$per}}</a>
                     @endforeach
                 </div>
             </div>
             <div class="col-sm">
                 <div
-                    style="text-align: center">{{__('admin.showing_to_of_entries', ['show' => $from, 'to' => $to, 'of' => $paginator->total()])}}</div>
+                        style="text-align: center">{{__('admin.showing_to_of_entries', ['show' => $from, 'to' => $to, 'of' => $paginator->total()])}}</div>
             </div>
             <div class="col-sm d-none d-lg-block d-xl-block">
                 @if ($paginator->hasPages())
@@ -26,7 +26,8 @@
                             @else
                                 <li class="page-item">
                                     <a class="page-link"
-                                       href="{{ urlWithGet([$page_name => $paginator->currentPage() - 1]) }}" rel="prev"
+                                       href="{{ admin_url_with_get([$page_name => $paginator->currentPage() - 1]) }}"
+                                       rel="prev"
                                        aria-label="@lang('pagination.previous')">&lsaquo;</a>
                                 </li>
                             @endif
@@ -36,7 +37,7 @@
                                 {{-- "Three Dots" Separator --}}
                                 @if (is_string($element))
                                     <li class="page-item disabled" aria-disabled="true"><span
-                                            class="page-link">{{ $element }}</span></li>
+                                                class="page-link">{{ $element }}</span></li>
                                 @endif
 
                                 {{-- Array Of Links --}}
@@ -44,10 +45,10 @@
                                     @foreach ($element as $page => $url)
                                         @if ($page == $paginator->currentPage())
                                             <li class="page-item active" aria-current="page"><span
-                                                    class="page-link">{{ $page }}</span></li>
+                                                        class="page-link">{{ $page }}</span></li>
                                         @else
                                             <li class="page-item"><a class="page-link"
-                                                                     href="{{ urlWithGet([$page_name => $page]) }}">{{ $page }}</a>
+                                                                     href="{{ admin_url_with_get([$page_name => $page]) }}">{{ $page }}</a>
                                             </li>
                                         @endif
                                     @endforeach
@@ -58,11 +59,13 @@
                             @if ($paginator->hasMorePages())
                                 <li class="page-item">
                                     <a class="page-link"
-                                       href="{{ urlWithGet([$page_name => $paginator->currentPage() + 1]) }}" rel="next"
+                                       href="{{ admin_url_with_get([$page_name => $paginator->currentPage() + 1]) }}"
+                                       rel="next"
                                        aria-label="@lang('pagination.next')">&rsaquo;</a>
                                 </li>
                             @else
-                                <li class="page-item disabled" aria-disabled="true" aria-label="@lang('pagination.next')">
+                                <li class="page-item disabled" aria-disabled="true"
+                                    aria-label="@lang('pagination.next')">
                                     <span class="page-link" aria-hidden="true">&rsaquo;</span>
                                 </li>
                             @endif
@@ -84,7 +87,7 @@
                             @else
                                 <li class="page-item">
                                     <a class="page-link"
-                                       href="{{ urlWithGet([$page_name => $paginator->currentPage() - 1]) }}"
+                                       href="{{ admin_url_with_get([$page_name => $paginator->currentPage() - 1]) }}"
                                        rel="prev">@lang('pagination.previous')</a>
                                 </li>
                             @endif
@@ -93,7 +96,7 @@
                             @if ($paginator->hasMorePages())
                                 <li class="page-item">
                                     <a class="page-link"
-                                       href="{{ urlWithGet([$page_name => $paginator->currentPage() + 1]) }}"
+                                       href="{{ admin_url_with_get([$page_name => $paginator->currentPage() + 1]) }}"
                                        rel="next">@lang('pagination.next')</a>
                                 </li>
                             @else